Move api functions to api file
