+++ src/redux/API.js
@@ -1,3 +1,5 @@
+import { addBook, removeBook, loadBooks } from './books/books';
+
 //! For get app id
 // const URL = 'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps';
 // useEffect(() => {
@@ -15,10 +17,18 @@
 
 const BOOK_STORE_URL = 'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/naN2smxjCVu4fr6NtvX5/books';
 //! Get data
-const getBooks = async () => {
-  const response = await fetch(BOOK_STORE_URL);
+// const getBooks = async () => {
+//   const response = await fetch(BOOK_STORE_URL);
+//   const data = await response.json();
+//   return data;
+// };
+
+export const getBooks = () => async (dispatch) => {
+  const response = await fetch(
+    'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/naN2smxjCVu4fr6NtvX5/books',
+  );
   const data = await response.json();
-  return data;
+  dispatch(loadBooks(data));
 };
 
 //! Post data
@@ -32,10 +42,18 @@ const postBook = async (book) => {
   return response;
 };
 
+export const addBookToApi = (book) => async (dispatch) => {
+  postBook(book);
+  dispatch(addBook(book));
+};
+
 //! Delete data
 const deleteBook = async (id) => {
   const URL = `${BOOK_STORE_URL}/${id}`;
   return fetch(URL, { method: 'DELETE' });
 };
 
-export { getBooks, postBook, deleteBook };
+export const removeBookFromApi = (id) => async (dispatch) => {
+  deleteBook(id);
+  dispatch(removeBook(id));
+};
