+++ src/redux/books/books.js
@@ -1,5 +1,3 @@
-import { deleteBook, postBook } from '../API';
-
 const ADD_BOOK = 'bookStore/books/ADD_BOOK';
 const REMOVE_BOOK = 'bookStore/books/REMOVE_BOOK';
 const LOAD_BOOKS = 'bookStore/books/LOAD_BOOKS';
@@ -36,27 +34,9 @@ export const removeBook = (id) => ({
   },
 });
 
-export const addBookToApi = (book) => async (dispatch) => {
-  postBook(book);
-  dispatch(addBook(book));
-};
-
-export const removeBookFromApi = (id) => async (dispatch) => {
-  deleteBook(id);
-  dispatch(removeBook(id));
-};
-
-const loadBooks = (apiState) => ({
+export const loadBooks = (apiState) => ({
   type: LOAD_BOOKS,
   payload: apiState,
 });
-
-export const getBooks = () => async (dispatch) => {
-  const response = await fetch(
-    'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi/apps/naN2smxjCVu4fr6NtvX5/books',
-  );
-  const data = await response.json();
-  dispatch(loadBooks(data));
-};
 
 export default reducer;
